[WiP/draft] Sub-domain world generators (#724)
* Initial sub-domain world generator push

* Class structure optimization

* Fixed urban generator; url for textures not found during runtime

* Build tunnel_generator; SystemPath.cc:357 runtime error finding file with URL; mesh orientation incorrect

* Removed old code

* Updated cmake and moved files to match new apps directory
